+++ python_plus/python_plus/scripts/list_requirements.py
@@ -24,7 +24,7 @@ except ImportError:
     import z0lib
 
 
-__version__ = "2.0.5"
+__version__ = "2.0.6"
 python_version = "%s.%s" % (sys.version_info[0], sys.version_info[1])
 
 #