+++ wok_code/wok_code/cvt_csv_2_xml.py
@@ -44,7 +44,7 @@ except ImportError:
     import z0lib
 
 
-__version__ = "2.0.5"
+__version__ = "2.0.6"
 
 msg_time = time.time()