+++ wok_code/wok_code/scripts/do_migrate.py
@@ -7,7 +7,7 @@ import re
 import lxml.etree as ET
 from python_plus import _b
 
-__version__ = "2.0.5"
+__version__ = "2.0.6"
 
 MIGRATE_TO_NEW = [
     (