+++ wok_code/wok_code/scripts/odoo_translation.py
@@ -14,7 +14,7 @@ from openpyxl import load_workbook, Workbook
 
 # from python_plus import unicodes
 
-__version__ = "2.0.5"
+__version__ = "2.0.6"
 
 
 MODULE_SEP = "\ufffa"