+++ wok_code/wok_code/scripts/please_z0bug.py
@@ -2,7 +2,7 @@
 # -*- coding: utf-8 -*-
 
 
-__version__ = "2.0.5"
+__version__ = "2.0.6"
 
 
 class PleaseZ0bug(object):