+++ wok_code/wok_code/ssh.py
@@ -3,7 +3,7 @@
 import os
 import sys
 
-__version__ = '2.0.5'
+__version__ = '2.0.6'
 
 
 def get_remote_user():