+++ zerobug/zerobug/tests/zerobug_test_03.py
@@ -12,7 +12,7 @@ import sys
 from zerobug import z0test
 
 
-__version__ = "2.0.4"
+__version__ = "2.0.5"
 
 MODULE_ID = 'zerobug'
 TEST_FAILED = 1