+++ zerobug/zerobug/tests/zerobug_test_05.py
@@ -10,7 +10,7 @@ import os
 import sys
 from zerobug import z0test
 
-__version__ = "2.0.4"
+__version__ = "2.0.5"
 
 MODULE_ID = 'zerobug'
 TEST_FAILED = 1